refactored application_card and created constants for mapping application status and tag colors
